feat(Core): class restructure
