+++ Runtime/CVRFury/SubComponents/CVRFuryMenuStore.cs
@@ -2,6 +2,7 @@ using System;
 using System.Collections.Generic;
 using UnityEngine;
 using static uk.novavoidhowl.dev.cvrfury.supporting_classes.runtime.General;
+using uk.novavoidhowl.dev.cvrfury.packagecore;
 
 namespace uk.novavoidhowl.dev.cvrfury.runtime
 {
