+++ Runtime/CVRFury/SubComponents/CVRFuryModule.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using UnityEngine;
 using uk.novavoidhowl.dev.vrcstub;
 using uk.novavoidhowl.dev.cvrfury.supporting_classes.runtime;
+using uk.novavoidhowl.dev.cvrfury.packagecore;
 using System.Reflection;
 
 namespace uk.novavoidhowl.dev.cvrfury.runtime
