+++ package.json
@@ -21,5 +21,5 @@
   },
   "unity": "2021.3",
   "unityRelease": "23f1",
-  "version": "0.136.1-experimental"
+  "version": "0.137.0-experimental"
 }
