+++ packagecore/classes.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: e8d5e6f814984294caceedf93b024f6c
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
