+++ packagecore/classes/CVRFuryModuleSupport.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 944204e43595f2b4a818d841a7ef04f2
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
