+++ packagecore/classes/CVRMirror.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 6c49e38d9e0d8684aa3f10d9c548d4da
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
