+++ packagecore/classes/CVRMirror/ParameterStreamEntry.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: e1776006342ec7445b06af43efa7faf1
+guid: f4985bc8bd2ea7f4eb6d9ffc0c05a0ba
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
